temp/pyproject.toml
[tool.poetry]
name = "pilot-project"
version = "0.1.0"
description = "Monorepo for business dashboard services"
authors = ["h95fit <h95fit@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
streamlit = "^1.28.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
business-backend = "services.business.backend.app.main:app"
business-frontend = "services.business.frontend.app:main"